LPD-37754 Separate tests related to permissions and schedule (FF LPD-15596) to a new file
